refactor(about-me): moved components to middle-content
