+++ src/app/about-me/middle-content/personal-project-list/personal-project-list.component.scss
@@ -1,0 +1,1 @@
+@import '../../about-me.component.scss';
