+++ src/app/about-me/middle-content/personal-project-list/personal-project.config.ts
@@ -1,4 +1,4 @@
-import { Project } from "../about-me.model";
+import { Project } from "../../about-me.model";
 
 
 export const personalProjectList: Project[] = [
