+++ src/app/about-me/middle-content/personal-project-list/personal-project/personal-project.component.scss
@@ -1,4 +1,4 @@
-@import '../../about-me.component.scss';
+@import '../../../about-me.component.scss';
 
 .project {
     clear: left;
