+++ src/app/about-me/middle-content/skills/skills.component.scss
@@ -1,4 +1,4 @@
-@import '../about-me.component.scss';
+@import '../../about-me.component.scss';
 
 .skills {
     padding-bottom: 20px;
