+++ src/app/about-me/middle-content/skills/skills.config.ts
@@ -1,4 +1,4 @@
-import { Skill } from "../about-me.model";
+import { Skill } from "../../about-me.model";
 
 export const skillsConfig: Skill[] = [
     {
