+++ src/app/about-me/middle-content/work-experience-list/work-experience-list.component.scss
@@ -1,4 +1,4 @@
-@import '../about-me.component.scss';
+@import '../../about-me.component.scss';
 
 .work-experience-title {
     margin-bottom: 50px;
