+++ src/app/about-me/middle-content/work-experience-list/work-experience-list.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { WorkExperience } from '../about-me.model';
+import { WorkExperience } from '../../about-me.model';
 import { workExperienceList } from './work-experience.config';
 
 @Component({
