+++ src/app/about-me/middle-content/work-experience-list/work-experience.config.ts
@@ -1,4 +1,4 @@
-import { WorkExperience } from "../about-me.model";
+import { WorkExperience } from "../../about-me.model";
 
 export const workExperienceList: WorkExperience[] = [
     {
