+++ src/app/about-me/middle-content/work-experience-list/work-experience/work-experience.component.scss
@@ -1,4 +1,4 @@
-@import '../../about-me.component.scss';
+@import '../../../about-me.component.scss';
 
 .workExperience-row {
     clear: left;
