+++ src/app/about-me/middle-content/work-experience-list/work-experience/work-experience.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { WorkExperience } from '../../about-me.model';
+import { WorkExperience } from '../../../about-me.model';
 
 @Component({
   selector: 'app-work-experience',
