--- src/app/about-me/personal-project-list/personal-project-list.component.scss
@@ -1,1 +1,0 @@
-@import '../about-me.component.scss';
